vendor: bump Pebble to cfd19d33cdca
cfd19d33 db: compute file count and size metrics incrementally
e34d759c internal/rangedel: format keys in Fragmenter panics
06a3fb73 internal/mkbench: fix dropped walkfunc error
61c629af internal/manifest: unexport LevelMetadata files
8fdce959 vfs: Add FS.GetFreeSpace, IsNoSpaceError
24bd87a5 db: fix tombstone elision, grandparent limit bug
a9d18a23 *: Update comment on Ingest() on callers needing to Sync()
42cb6336 internal/manifest: take a LevelSlice in L0Sublevels.PickBaseCompaction
887d40ba db: use manifest.NewVersion constructor in tests
2f3fb56c internal/rangedel: Skip range tombstones outside a file's bounds
1a5f7680 record: propagate WAL sync errors to queued writers
cc76c76e db: update L0 compaction-picking to use manifest.LevelIterator
69961893 pebble: fix dropped test error
5edc4f5b db: use LevelSlice.SizeSum where applicable
d524d342 db: use manifest.LevelIterator in compaction, levelIter
049c566a db: capture stack trace of closed call for ErrClosed

Release note: None

+++ pkg/storage/pebble.go
@@ -819,7 +819,7 @@ func (p *Pebble) GetEnvStats() (*EnvStats, error) {
 	stats.TotalBytes = m.WAL.Size + m.Table.ZombieSize
 	for _, l := range m.Levels {
 		stats.TotalFiles += uint64(l.NumFiles)
-		stats.TotalBytes += l.Size
+		stats.TotalBytes += uint64(l.Size)
 	}
 
 	sstSizes := make(map[pebble.FileNum]uint64)
